requirements/app.txt
galaxy.plugin.api==0.47
requests==2.21.0
psutil==5.6.1
pyobjc-core==5.2; sys_platform == 'darwin'
pyobjc-framework-Cocoa==5.2; sys_platform == 'darwin'
pyobjc-framework-Quartz==5.2; sys_platform == 'darwin'
protobuf==3.7.1
fnvhash==0.1.0